+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - API
     - Devices
       - Fields
-        - new Features
+        - connectionType
+        - new OptionalFeatures
     - Home
       - accessPointUpdateStates
 ### CHANGED