+++ CHANGELOG.md
@@ -5,7 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### ADDED
+- API
+    - Devices
+      - HMIPW-DRD3 (Homematic IP Wired Dimming Actuator – 3x channels)
+      - HMIPW-DRS8 (Homematic IP Wired Switch Actuator – 8x channels)
+      - HMIPW-DRI32 (Homematic IP Wired Inbound module – 32x channels)
+      - HMIP-STV (Inclination and vibration Sensor)
+      - Fields
+        - connectionType
+        - new OptionalFeatures
+    - Home
+      - accessPointUpdateStates
 ### CHANGED
 - [BUG: 325] Requirements are now using a min version instead of a pinned version. requirements_dev.txt will still use the pinned versions to make sure that the latest version is compatible with the library.